Update README.md
added pictures and removed the challenges section: "Challenges and Difficulties Encountered"

I encountered difficulty with the “countifs” function.  To use the formula, I searched for information on “countifs” and applied those processes to the Excel.  My initial chart departed from the example in the module.  By chatting with a classmate, I realized that I forgot to include the “greater than”(>) symbol in the formula which skewed the table. Once I corrected the formula, the chart matched the example in the module.